+++ CHANGELOG.md
@@ -36,7 +36,7 @@
     - air-primary, air-primary-success, air-primary-alert, air-primary-copilot
     - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2, air-secondary-no-accent
     - air-tertiary, air-tertiary-accent, air-tertiary-no-accent
-    - air-selection
+    - air-selection, air-boost
     - @deprecate
       - default, danger, success, warning, primary, secondary, collab, ai, link
   - depth - @deprecate